+++ crates/swc_ecma_magicast/src/option.rs
@@ -1,0 +1,12 @@
+use crate::Proxy;
+
+pub struct OptionalNode<T>
+where
+    T: Proxy(Option<T>);
+
+impl<T> Proxy for OptionalNode<T>
+where
+    T: Proxy,
+{
+    type Item = Option<T::Item>;
+}
